Implement storage config resolution for shared tables and improve flush job tracking
SharedTableService now resolves storage location using storage configuration and path templates, ensuring correct directory structure for shared tables. Flush jobs update their status in the jobs table upon completion or failure, improving observability and reliability. Additional minor improvements include support for Float64 in JSON batch builder, code formatting, and test enhancements.

+++ backend/tests/integration/common/fixtures.rs
@@ -197,7 +197,7 @@ pub async fn create_shared_table(
             conversation_id TEXT NOT NULL,
             title TEXT,
             status TEXT,
-            participant_count INTEGER,
+            participant_count BIGINT,
             created_at TIMESTAMP
         "#
     };
